# Conflicts:
#	fannie/cron/nightly.clipboard.php
#	pos/is4c-nf/ajax-callbacks/ajax-end.php
#	pos/is4c-nf/gui-modules/HouseCouponList.php
#	pos/is4c-nf/gui-modules/adminlogin.php
#	pos/is4c-nf/gui-modules/login3.php
#	pos/is4c-nf/gui-modules/mgrlogin.php
#	pos/is4c-nf/gui-modules/refundComment.php
#	pos/is4c-nf/install/InstallUtilities.php
#	pos/is4c-nf/install/receipt.php
#	pos/is4c-nf/install/security.php
#	pos/is4c-nf/lib/MiscLib.php
#	pos/is4c-nf/lib/ReceiptLib.php
#	pos/is4c-nf/parser-class-lib/parse/UPC.php
#	pos/is4c-nf/plugins/Paycards/FirstData.php
#	pos/is4c-nf/plugins/QuickMenus/quickmenus/10.php
#	pos/is4c-nf/plugins/QuickMenus/quickmenus/9.php
#	pos/is4c-nf/scale-drivers/drivers/NewMagellan/ports.conf

+++ fannie/cron/nightly.specialorder.php
@@ -31,21 +31,27 @@
 
 */
 
-include('../config.php');
-include($FANNIE_ROOT.'src/SQLManager.php');
-include($FANNIE_ROOT.'src/cron_msg.php');
+include(dirname(__FILE__) . '/../config.php');
+if (!class_exists('FannieAPI')) {
+    include($FANNIE_ROOT . 'classlib2.0/FannieAPI.php');
+}
+if (!function_exists('cron_msg')) {
+    include($FANNIE_ROOT.'src/cron_msg.php');
+}
 
 set_time_limit(0);
 
 // clean cache
 $cachepath = sys_get_temp_dir()."/ordercache/";
-$dh = opendir($cachepath);
-while (($file = readdir($dh)) !== false) {
-    if ($file == "." || $file == "..") continue;
-    if (!is_file($cachepath.$file)) continue;
-    unlink($cachepath.$file);
+if (file_exists($cachepath)) {
+    $dh = opendir($cachepath);
+    while (($file = readdir($dh)) !== false) {
+        if ($file == "." || $file == "..") continue;
+        if (!is_file($cachepath.$file)) continue;
+        unlink($cachepath.$file);
+    }
+    closedir($dh);
 }
-closedir($dh);
 
 $sql = new SQLManager($FANNIE_SERVER,$FANNIE_SERVER_DBMS,$FANNIE_TRANS_DB,
         $FANNIE_SERVER_USER,$FANNIE_SERVER_PW);
@@ -253,17 +259,3 @@ if (strlen($empty) > 2){
     $delR = $sql->query($delQ);
 }
 
-/* blueLine flagging disabled
-$q = "SELECT card_no,count(*) FROM PendingSpecialOrder as p LEFT JOIN
-    SpecialOrders AS s ON p.order_id=s.specialOrderID
-    WHERE s.statusFlag=5 AND trans_id > 0 
-    GROUP BY card_no";
-$r = $sql->query($q);
-while($w = $sql->fetch_row($r)){
-    $upQ = "UPDATE {$FANNIE_OP_DB}.custdata SET blueLine=CONCAT(blueLine, ' SO({$w[1]})')
-        WHERE CardNo={$w[0]}";
-    $upR = $sql->query($upQ);
-}
-*/
-
-?>
